requirements.txt
Flask
Flask-JWT-Extended
Flask-PyMongo
requests
python-dotenv
